# GPU gemv->dot speedup for new backend