--- a/theano/gpuarray/blas.py
+++ b/theano/gpuarray/blas.py
@@ -138,7 +138,7 @@ class GpuGemv(BlasOp):
         return code
 
     def c_code_cache_version(self):
-        return (5,)
+        return (6,)
 
 gpugemv_no_inplace = GpuGemv(inplace=False)
 gpugemv_inplace = GpuGemv(inplace=True)