Similar to #5257, this PR applies to the new backend.

~~**ONLY merge this after libgpuarray-294 is merged.**~~
To test this, newer `libgpuarray` with [this PR](https://github.com/Theano/libgpuarray/pull/294) merged is needed.

Results of running [this](https://gist.github.com/khaotik/9c7c72e562a919e241c9a09522fc4df5) test script on my laptop (GeForce GT 650m):

(time used in seconds, lower is better)

Before:
```
  scan with dot  |  scan with sum   |  full unrolled
 3.502193689346  |  0.411211490631  |  0.089465141296  
 1.447018623352  |  0.412187576294  |  0.086945295334  
 1.446414470673  |  0.411444425583  |  0.089369773865  
 1.447574138641  |  0.413084506989  |  0.091281890869  
 1.452363491058  |  0.412972927094  |  0.088017463684  
...
```

After:
```
  scan with dot  |  scan with sum   |  full unrolled
 2.389663457870  |  0.413246393204  |  0.089257478714  
 0.485583782196  |  0.413187980652  |  0.086747169495  
 0.486212968826  |  0.413235425949  |  0.088805198669  
 0.489252328873  |  0.413118600845  |  0.088519096375  
 0.484216928482  |  0.413124799728  |  0.087159156799  
...
```

I'm not sure this also works for OpenCL. Could use some help for testing. For OpenCL device, both Intel/AMD GPU, both clBLAS and clBLAST library, should be tested.